Improve instrumentation
Do not rely on static analysis of what function can be spawned in
threads. Change thrd_create to call our wrapper that takes care
of all the setup/tear down of our structures (except for creating
the subbuffer that we still need to do from the calling thread
to avoid locking of the parent buffer).